fix: move quiz to model package
