+++ app/src/main/java/com/ottf/quizdroid/presentation/quiz/components/AnswerSheet.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.ottf.quizdroid.domain.Quiz
+import com.ottf.quizdroid.domain.model.Quiz
 import com.ottf.quizdroid.ui.theme.CustomTypography
 
 @Composable
